feat(ms_solicitudes): Refactor user data handling and introduce ValidateUserUseCase
