+++ ms_solicitudes/applications/app-service/src/main/java/co/com/leronarenwino/config/UseCasesConfig.java
@@ -1,6 +1,7 @@
 package co.com.leronarenwino.config;
 
 import co.com.leronarenwino.model.gateway.LoanApplicationRepository;
+import co.com.leronarenwino.model.gateway.RestConsumerService;
 import co.com.leronarenwino.model.gateway.SenderService;
 import co.com.leronarenwino.usecase.*;
 import org.springframework.context.annotation.*;
@@ -32,6 +33,13 @@ public class UseCasesConfig {
     public SendNotificationUseCase sendNotificationUseCase(
             SenderService senderService) {
         return new SendNotificationUseCase(senderService);
+    }
+
+    @Bean
+    @Primary
+    public ValidateUserUseCase validateUserUseCase(
+            RestConsumerService restConsumerService) {
+        return new ValidateUserUseCase(restConsumerService);
     }
 
     @Bean
