+++ ms_solicitudes/infrastructure/driven-adapters/rest-consumer/src/main/java/co/com/leronarenwino/consumer/dto/GenericResponse.java
@@ -1,12 +1,14 @@
 package co.com.leronarenwino.consumer.dto;
 
+import co.com.leronarenwino.model.UserData;
+
 public record GenericResponse<T>(
         String message,
         T data,
         String timestamp,
         int status
 ) {
-    public static UserResponse toUserResponse(GenericResponse<UserResponse> genericResponse) {
+    public static UserData toUser(GenericResponse<UserData> genericResponse) {
         return genericResponse.data();
     }
 }
