+++ ms_solicitudes/infrastructure/entry-points/reactive-web/src/main/java/co/com/leronarenwino/api/dto/UserDataResponse.java
@@ -1,12 +1,12 @@
 package co.com.leronarenwino.api.dto;
 
-import co.com.leronarenwino.consumer.dto.UserResponse;
+import co.com.leronarenwino.model.UserData;
 
 public record UserDataResponse(
         String name,
         Double baseSalary
 ) {
-    public static UserDataResponse toUserData(UserResponse userData) {
+    public static UserDataResponse toUserDataResponse(UserData userData) {
         return new UserDataResponse(
                 userData.name(),
                 userData.baseSalary()
